CineQuery-master/requirements.txt
streamlit==1.35.0
google-cloud-bigquery==3.21.0
google-auth==2.29.0
requests==2.31.0
db-dtypes==1.2.0
